server/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.25
pymysql==1.1.0
alembic==1.13.1
cryptography==41.0.7

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.8.0

# Pydantic for validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Redis for caching and sessions
redis==5.0.1
hiredis==2.3.2

# Payment Providers
stripe==7.10.0
requests==2.31.0

# Image Processing
Pillow==10.2.0
python-magic==0.4.27

# AWS S3 for file storage
boto3==1.34.34

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# WebSocket support
websockets==12.0

# Background tasks
celery==5.3.6
flower==2.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
phonenumbers==8.13.27

# Location and GIS
geopy==2.4.1
shapely==2.0.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Monitoring and Logging
python-json-logger==2.0.7

# Environment Variables
python-dotenv==1.0.0

# CORS
fastapi-cors==0.0.6

# Rate limiting
slowapi==0.1.9

# Excel file processing
openpyxl==3.1.2